requirements.txt
fastapi
opencv-python
mediapipe
pyttsx3
passlib
PyJWT
python-dotenv
python-multipart
uvicorn
uvicorn[standard]
websockets
wsproto
bcrypt